# storage/engine: implement iterate upper-bound for batch iterators