Revenge of the deletion tombstones. This was the same problem as
but that fix did not apply to batch iterators.

This doesn't fix the Delete100 performance yet as EndTransaction is
using an iterator without an upper-bound. And for some unknown reason,
switching to an iterator with an upper-bound doesn't help.

```
name                old time/op  new time/op  delta
IterOnBatch10-8     3.99µs ± 1%  2.57µs ± 2%  -35.49%  (p=0.000 n=10+10)
IterOnBatch100-8    26.8µs ± 1%   3.0µs ± 1%  -88.82%   (p=0.000 n=10+8)
IterOnBatch1000-8    405µs ± 0%     4µs ±10%  -99.04%   (p=0.000 n=8+10)
IterOnBatch10000-8  5.26ms ± 1%  0.00ms ± 1%  -99.92%  (p=0.000 n=10+10)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/5332)
<!-- Reviewable:end -->