@tamird 2016-03-17
> This was the same problem as but that fix did not apply to batch iterators.

missing a word

---

Reviewed 1 of 2 files at r1.
Review status: 1 of 2 files reviewed at latest revision, all discussions resolved.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/5332#-:-KD4bpwPs_1shEYPvfmG:-453415962)*
<!-- Sent from Reviewable.io -->