@dt 2016-03-17
Reviewed 1 of 2 files at r1.
Review status: all files reviewed at latest revision, all discussions resolved.

---



*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/5332)*
<!-- Sent from Reviewable.io -->