@petermattis 2016-03-17
Updated the commit message.